backend/telegram/cheeta-assistant-bot/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
pyaes==1.6.1
Pyrogram==1.4.8
PySocks==1.7.1
python-dotenv==0.19.2
requests==2.27.1
TgCrypto==1.2.3
urllib3==1.26.8
